Move all compat functions/imports over to common.py
